Prepare for implementing service
