+++ src/server.h
@@ -1,0 +1,8 @@
+#ifndef __SERVER_H__
+#define __SERVER_H__
+
+#include <windef.h>
+
+int  iServerLoop(BOOL);
+
+#endif /* __SERVER_H__ */
